refactor: removed () in folder names of components in question-type
