+++ app/questionnaire/components/QuestionCard.tsx
@@ -6,11 +6,11 @@ import { useParams, useRouter } from "next/navigation";
 import { useCallback } from "react";
 
 import { useQuestionContext } from "../context/question-context";
-import MultipleChoiceQuestion from "./(question-types)/(multiple-choice-question)";
-import SingleChoiceQuestion from "./(question-types)/(single-choice-question)";
-import TextQuestion from "./(question-types)/(text-question)";
 import InvalidComponent from "./InvalidComponent";
 import SubmitAlert from "./SubmitAlert";
+import MultipleChoiceQuestion from "./question-types/multiple-choice-question";
+import SingleChoiceQuestion from "./question-types/single-choice-question";
+import TextQuestion from "./question-types/text-question";
 
 const QUESTION_COMPONENTS: Record<QuestionTypeValues, React.FC<any>> = {
   text: TextQuestion,
@@ -39,7 +39,9 @@ const QuestionCard = () => {
     [saveAnswer],
   );
 
-  const QuestionComponent = QUESTION_COMPONENTS[currentQuestion?.type];
+  const QuestionComponent = currentQuestion?.type
+    ? QUESTION_COMPONENTS[currentQuestion.type as QuestionTypeValues]
+    : null;
 
   if (!currentQuestion) {
     return <div>No question found</div>;
